[UP] added new print_cmd

+++ include/needforstek.h
@@ -22,6 +22,6 @@ typedef struct info_s
 info_t *init_struct(void);
 int print_cmd(info_t *info, char *str);
 int ai(info_t *info);
-int print_float_cmd(info_t *info, char *str, double val);
+int print_float_cmd(info_t *info, char *str, float val);
 
 #endif /* !MY_PROG_H_ */
